the ai now gets closer to it's ennemy

+++ src/main/java/com/glhf/bomberball/ai/AlphaBetaReturnObj.java
@@ -5,12 +5,12 @@ import com.glhf.bomberball.utils.Action;
 import java.util.List;
 
 public class AlphaBetaReturnObj {
-    public int score;
+    public double score;
     public List<Action> actions;
     public List<List<Action>> actionsPossibles;
     public int badness;
 
-    AlphaBetaReturnObj(int score, List<Action> actions,List<List<Action>> actionsPossibles,int badness){
+    AlphaBetaReturnObj(double score, List<Action> actions,List<List<Action>> actionsPossibles,int badness){
         this.score=score;
         this.actions=actions;
         this.actionsPossibles=actionsPossibles;
